Rename thomas Styles.js
