+++ src/components/ChemicalModels/Thomas/ThomasStyles.js
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import styled from "styled-components/macro";
 
 export const ThomasPageLayout = styled.div`
   display: grid;
